Add files via upload
The primary goal of this model is to determine how well our machine learning algorithm can predict and
train our data, as well as how increasing in number of epochs and data size will enhance our model. When
compared to other stock price prediction model, this will give more accurate results. The accuracy of this
model is tested with various data sizes and epochs, and the results are tabulated. This article mainly focuses
on developing a model for predicting future stock market values using Recurrent Neural Networks (RNN)
and, in particular, the Long-Short Term Memory model (LSTM).